Include wasm build files in cache key

+++ sqlite3_wasm_build/src/external_objects.c
@@ -10,20 +10,22 @@ static __externref_t host_objects[0];
 
 // Note: Inlining these functions seems to cause linker errors (something about
 // clang attempting to take the address of host_objects).
-void host_table_set(int index, __externref_t value) {
-  __builtin_wasm_table_set(host_objects, index, value);
+void host_table_set(size_t index, __externref_t value) {
+  __builtin_wasm_table_set(host_objects, (int) index, value);
 }
 
-static __externref_t host_table_get(int index) {
-  return __builtin_wasm_table_get(host_objects, index);
+static __externref_t host_table_get(size_t index) {
+  return __builtin_wasm_table_get(host_objects, (int) index);
 }
 
-static int host_table_grow(int delta) {
-  return __builtin_wasm_table_grow(host_objects,
-                                   __builtin_wasm_ref_null_extern(), delta);
+static size_t host_table_grow(size_t delta) {
+  return (size_t)__builtin_wasm_table_grow(
+      host_objects, __builtin_wasm_ref_null_extern(), delta);
 }
 
-static int host_table_size() { return __builtin_wasm_table_size(host_objects); }
+static size_t host_table_size() {
+  return (size_t)__builtin_wasm_table_size(host_objects);
+}
 
 // A simple slab allocator to find free indices in the host objects table.
 typedef struct {
